bash script v 0.2
This is a bash script to quickly and easily toggle Snapper's automatic features.

Use it, for example, while you are working to recover from an install that went wrong and
when you temporarily don't want to risk accidentally loosing valuable snapshots that
otherwise might be lost as part of the automatic clean up process.  Especially useful when a
late Friday crash occurs, and you want to be able to safely shutdown and then start the 
recovery process on Monday morning.

It both disables the automatic deletion of older snapshots (i.e. it disables automatic 
'CLEANUPs') and disables the creation of new hourly snapshots.  In other words it moves
Snapper into, and then out of, a quiescent state, and in all partitions that it's configured for.

When run a second time, it restores Snapper to exactly how it was previously configured.